+++ preload.cjs
@@ -1,6 +1,7 @@
 const { contextBridge, ipcRenderer } = require('electron');
 
 contextBridge.exposeInMainWorld('blip', {
+  platform: process.platform,
   getConfig: () => ipcRenderer.invoke('get-config'),
   saveConfig: (config) => ipcRenderer.invoke('save-config', config),
   getAppMetadata: () => ipcRenderer.invoke('get-app-metadata'),